+++ packages/deepsea-tools/src/utils/isNullable.ts
@@ -1,0 +1,5 @@
+import { isNonNullable } from "./isNonNullable"
+
+export function isNullable<T>(value: T): value is Exclude<T, NonNullable<T>> {
+    return !isNonNullable(value)
+}